embellish: add fenceposts, infer side as rotated top
Many repeating border patterns have a fencepost-and-fence style of
repetition, where there is some "overlapping" element between
iterations. Support this with an optional "fencepost" argument.

Remove support for distinct `side` patterns, and from now on assume
that all borders are rotated versions of the top. (it was annoying to
think about specifying two fenceposts, and all my test patterns have a
side.pattern that's just a rotated top.pattern that I have to make
separately for no good reason)